requirements.txt
earthengine-api
google-cloud-storage
numpy
rasterio
requests
